DatadogAgentClient

Documentation for DatadogAgentClient.

DatadogAgentClient.DatadogAgent

DatadogAgentClient.count

DatadogAgentClient.decrement

DatadogAgentClient.distribution

DatadogAgentClient.gauge

DatadogAgentClient.histogram

DatadogAgentClient.increment

DatadogAgentClient.set

DatadogAgentClient.timing

DatadogAgentClient.DatadogAgent — Method

DatadogAgent(ip::String="127.0.0.1", port=8125)
DatadogAgent(ip::IPAddr, port)

Create a client to send metrics to a local Datadog agent at the given ip and port. Should be created only once and used for all metrics.

Examples

julia> client = DatadogAgent()
julia> increment(client, "visits", "test:true", "env:prod")

DatadogAgentClient.count — Method

count(cl::DatadogAgent, metric, value, tags::String...)

Increment the counter metric by value, optionally with tags.

Examples

julia> count(DatadogAgent(), "items_sold", 42)
julia> count(DatadogAgent(), "items_sold", 42, "source:homepage")

DatadogAgentClient.decrement — Method

decrement(cl::DatadogAgent, metric, tags::String...)

Decrement the counter metric by 1, optionally with tags.

Examples

julia> decrement(DatadogAgent(), "available_seats")
julia> decrement(DatadogAgent(), "available_seats", "env:prod")

DatadogAgentClient.distribution — Method

distribution(cl::DatadogAgent, metric, value, tags::String...)

Adds value to distribution metric, optionally with tags.

Examples

julia> distribution(DatadogAgent(), "response_time", 25)
julia> distribution(DatadogAgent(), "response_time", 25, "service:reporting")

DatadogAgentClient.gauge — Method

gauge(cl::DatadogAgent, metric, value, tags::String...)

Set the gauge metric to value, optionally with tags.

Examples

julia> gauge(DatadogAgent(), "cpu_load", 3.9)
julia> gauge(DatadogAgent(), "cpu_load", 3.9, "instance:dev")

DatadogAgentClient.histogram — Method

histogram(cl::DatadogAgent, metric, value, tags::String...)

Adds value to histogram metric, optionally with tags.

Examples

julia> histogram(DatadogAgent(), "shopping_cart_size", 5)
julia> histogram(DatadogAgent(), "shopping_cart_size", 5, "promotion:end_of_year")

DatadogAgentClient.increment — Method

increment(cl::DatadogAgent, metric, tags::String...)

Increment the counter metric by 1, optionally with tags.

Examples

julia> increment(DatadogAgent(), "visits")
julia> increment(DatadogAgent(), "visits", "test:true")

DatadogAgentClient.set — Method

set(cl::DatadogAgent, metric, value, tags::String...)

Adds value to set metric, optionally with tags.

Examples

julia> set(DatadogAgent(), "unique_user", userId)
julia> set(DatadogAgent(), "unique_user", userId, "foo:bar")

DatadogAgentClient.timing — Method

timing(cl::DatadogAgent, metric, value, tags::String...)

Send a timing value for metric, optionally with tags.

Examples

julia> timing(DatadogAgent(), "processing_time", 23.5)
julia> timing(DatadogAgent(), "processing_time", 23.5, "foo:bar")