DatadogClient

Documentation for DatadogClient.

DatadogClient.Datadog

DatadogClient.count

DatadogClient.decrement

DatadogClient.distribution

DatadogClient.gauge

DatadogClient.histogram

DatadogClient.increment

DatadogClient.set

DatadogClient.timing

DatadogClient.Datadog — Method

Datadog(ip::String="127.0.0.1", port=8125)
Datadog(ip::IPAddr, port)

Create a Datadog client sending metrics to the given ip and port. Should be created only once and used for all metrics.

Examples

julia> client = Datadog()
julia> increment(client, "visits", "test:true", "env:prod")

DatadogClient.count — Method

count(cl::Datadog, metric, value, tags::String...)

Increment the counter metric by value, optionally with tags.

Examples

julia> count(Datadog(), "items_sold", 42)
julia> count(Datadog(), "items_sold", 42, "source:homepage")

DatadogClient.decrement — Method

decrement(cl::Datadog, metric, tags::String...)

Decrement the counter metric by 1, optionally with tags.

Examples

julia> decrement(Datadog(), "available_seats")
julia> decrement(Datadog(), "available_seats", "env:prod")

DatadogClient.distribution — Method

distribution(cl::Datadog, metric, value, tags::String...)

Adds value to distribution metric, optionally with tags.

Examples

julia> distribution(Datadog(), "response_time", 25)
julia> distribution(Datadog(), "response_time", 25, "service:reporting")

DatadogClient.gauge — Method

gauge(cl::Datadog, metric, value, tags::String...)

Set the gauge metric to value, optionally with tags.

Examples

julia> gauge(Datadog(), "cpu_load", 3.9)
julia> gauge(Datadog(), "cpu_load", 3.9, "instance:dev")

DatadogClient.histogram — Method

histogram(cl::Datadog, metric, value, tags::String...)

Adds value to histogram metric, optionally with tags.

Examples

julia> histogram(Datadog(), "shopping_cart_size", 5)
julia> histogram(Datadog(), "shopping_cart_size", 5, "promotion:end_of_year")

DatadogClient.increment — Method

increment(cl::Datadog, metric, tags::String...)

Increment the counter metric by 1, optionally with tags.

Examples

julia> increment(Datadog(), "visits")
julia> increment(Datadog(), "visits", "test:true")

DatadogClient.set — Method

set(cl::Datadog, metric, value, tags::String...)

Adds value to set metric, optionally with tags.

Examples

julia> set(Datadog(), "unique_user", userId)
julia> set(Datadog(), "unique_user", userId, "foo:bar")

DatadogClient.timing — Method

timing(cl::Datadog, metric, value, tags::String...)

Send a timing value for metric, optionally with tags.

Examples

julia> timing(Datadog(), "processing_time", 23.5)
julia> timing(Datadog(), "processing_time", 23.5, "foo:bar")